semana5/requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozendict==2.4.4
greenlet==3.0.3
html5lib==1.1
idna==3.7
lxml==5.2.2
multitasking==0.0.11
numpy==1.26.4
packaging==24.0
pandas==2.0.3
peewee==3.17.5
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.52
sqlalchemy-redshift==0.8.14
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
yfinance==0.2.40
